Update: Redis Queue Lock

+++ api/v1/sms/sub_controller_sms_management.go
@@ -150,7 +150,7 @@ type SubPostConversationRecordRes struct {
 }
 
 type SubPostConversationRecordData struct {
-	TaskID            int    `json:"task_id"`
+	TaskID            int64  `json:"task_id"`
 	Content           string `json:"content"`
 	DeviceNumber      string `json:"device_number"`
 	TargetPhoneNumber string `json:"target_phone_number"`
